[camera] Refactor Camera/Android implementation to allow reuse of Camera functionality (#1949)
